requirements.txt
requests
colorama
beautifulsoup4
asyncio
statistics
aiohttp
speedtest-cli
PyInquirer
prompt_toolkit